# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EALifecycle API Documentation",
        "description": "API documentation for EALifecycle - The definitive platform for managing Expert Advisors professionally",
        "contact": {
            "email": "support@ealifecycle.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local Environment"
        }
    ],
    "paths": {
        "/api/v1/portfolios": {
            "get": {
                "tags": [
                    "Portfolios"
                ],
                "summary": "Get list of portfolios",
                "operationId": "fecb3366147bbc9da48af03db756f340",
                "responses": {
                    "200": {
                        "description": "List of portfolios"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Portfolios"
                ],
                "summary": "Create a new portfolio",
                "operationId": "b785e229770a8f14a04ac1ce8a1b2a0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "group_id": {
                                        "type": "integer"
                                    },
                                    "strategy_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Portfolio created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/portfolios/{id}": {
            "get": {
                "tags": [
                    "Portfolios"
                ],
                "summary": "Get a specific portfolio",
                "operationId": "ed9f7b2d7c0239dadacc11ede321f9ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Portfolio details"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Portfolios"
                ],
                "summary": "Update a portfolio",
                "operationId": "737889930932af68af36d115941cbf03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "group_id": {
                                        "type": "integer"
                                    },
                                    "strategy_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Portfolio updated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Portfolios"
                ],
                "summary": "Delete a portfolio",
                "operationId": "62d25e9fc72ffeb90f406a9899bd1dc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/strategies": {
            "get": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Get list of strategies",
                "operationId": "d116f14007fc505acd6b968c3b6648af",
                "responses": {
                    "200": {
                        "description": "List of strategies"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Create a new strategy",
                "operationId": "dd512a03b511d1976f017ff8a7302f23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "timeframe_ids",
                                    "primary_timeframe_id",
                                    "status_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "symbols_traded": {
                                        "type": "string"
                                    },
                                    "magic_number": {
                                        "type": "integer"
                                    },
                                    "timeframe_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "primary_timeframe_id": {
                                        "type": "integer"
                                    },
                                    "group_id": {
                                        "type": "integer"
                                    },
                                    "status_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Strategy created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/strategies/{id}": {
            "get": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Get a specific strategy",
                "operationId": "833865a097211671974f77fc55d682fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Strategy details"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Update a strategy",
                "operationId": "5448f1a30f62d74930a296c2ec2034c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "timeframe_ids",
                                    "primary_timeframe_id",
                                    "status_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "symbols_traded": {
                                        "type": "string"
                                    },
                                    "magic_number": {
                                        "type": "integer"
                                    },
                                    "timeframe_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "primary_timeframe_id": {
                                        "type": "integer"
                                    },
                                    "group_id": {
                                        "type": "integer"
                                    },
                                    "status_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Strategy updated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Delete a strategy",
                "operationId": "a556d384d629386e1b65616c0315c181",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/strategies/{id}/source-code": {
            "get": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Download source code for a strategy",
                "operationId": "3f47565a1e7a06f2ad7c27cbdfb1c16a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File download"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "File not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Strategies"
                ],
                "summary": "Upload source code for a strategy",
                "operationId": "313188d7dc5741f5d3e5084550926fa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "source_code_file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/strategies/{strategy}/reports": {
            "get": {
                "tags": [
                    "Strategy Reports"
                ],
                "summary": "Download the strategy's report",
                "description": "Download a specific report",
                "operationId": "36771693644fe071943f14c0fcf9858f",
                "parameters": [
                    {
                        "name": "strategy",
                        "in": "path",
                        "description": "Strategy ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file download",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Report not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Strategy Reports"
                ],
                "summary": "Upload a new PDF report",
                "description": "Upload a new PDF report for a strategy",
                "operationId": "189cac67740201956cb4cd3650000e12",
                "parameters": [
                    {
                        "name": "strategy",
                        "in": "path",
                        "description": "Strategy ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "report": {
                                        "description": "PDF report file (max 10MB)",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Report uploaded successfully"
                                        },
                                        "filename": {
                                            "type": "string"
                                        },
                                        "original_filename": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Strategy Reports"
                ],
                "summary": "Delete the strategy's report",
                "description": "Delete a report",
                "operationId": "ec4baf21d345711184792481c1a06c42",
                "parameters": [
                    {
                        "name": "strategy",
                        "in": "path",
                        "description": "Strategy ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Report deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Report not found"
                    }
                }
            }
        },
        "/api/v1/symbols": {
            "get": {
                "tags": [
                    "Symbols"
                ],
                "summary": "Get a list of symbols",
                "operationId": "7412fd3a8ff6f1cd4a055a7dbfb6147a",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of symbols",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SymbolResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Symbols"
                ],
                "summary": "Create a new symbol",
                "operationId": "6d300208a2d500e38a26125762bbbde0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "symbol"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "EURUSD"
                                    },
                                    "symbol": {
                                        "type": "string",
                                        "example": "Euro / US Dollar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Symbol created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SymbolResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/symbols/{id}": {
            "get": {
                "tags": [
                    "Symbols"
                ],
                "summary": "Get a specific symbol",
                "operationId": "8f941bd6a0a582581e8023ab0f43e219",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Symbol details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SymbolResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Symbol not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Symbols"
                ],
                "summary": "Update a symbol",
                "operationId": "1284e6f767b12cf286ba0dbd82ea6fe0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "symbol"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "symbol": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Symbol updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SymbolResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Symbol not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Symbols"
                ],
                "summary": "Delete a symbol",
                "operationId": "d6b2faac54e88440259969f381622889",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Symbol deleted successfully"
                    },
                    "404": {
                        "description": "Symbol not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "SymbolResource": {
                "title": "Symbol Resource",
                "description": "A trading symbol",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "EURUSD"
                    },
                    "symbol": {
                        "type": "string",
                        "example": "Euro / US Dollar"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Portfolios",
            "description": "API Endpoints for Managing Portfolios"
        },
        {
            "name": "Strategies",
            "description": "API Endpoints for Managing Strategies"
        },
        {
            "name": "Strategy Reports",
            "description": "API Endpoints for managing strategy PDF reports"
        },
        {
            "name": "Symbols",
            "description": "API Endpoints for Managing Trading Symbols"
        }
    ]
}